+++ ci/cpu/cuxfilter/build_cuxfilter.sh
@@ -1,4 +1,0 @@
-set -e
-
-echo "Building cuxfilter"
-conda build conda/recipes/cuxfilter --python=$PYTHON