+++ docs/source/conf.py
@@ -27,9 +27,9 @@ author = 'NVIDIA'
 # built documents.
 #
 # The short X.Y version.
-version = '0.16'
+version = '0.17'
 # The full version, including alpha/beta/rc tags
-release = '0.16.1'
+release = '0.17.0'
 
 nbsphinx_allow_errors = True
 # -- General configuration ---------------------------------------------------