+++ python/cuxfilter/charts/core/non_aggregate/core_scatter.py
@@ -1,6 +1,7 @@
 from typing import Tuple
 
 from .core_non_aggregate import BaseNonAggregate
+from ...constants import CUXF_DEFAULT_COLOR_PALETTE
 
 
 class BaseScatter(BaseNonAggregate):
@@ -22,7 +23,7 @@ class BaseScatter(BaseNonAggregate):
         x_range=None,
         y_range=None,
         add_interaction=True,
-        color_palette=None,
+        color_palette=CUXF_DEFAULT_COLOR_PALETTE,
         aggregate_col=None,
         aggregate_fn="count",
         point_size=1,
@@ -78,7 +79,8 @@ class BaseScatter(BaseNonAggregate):
         else:
             self.aggregate_col = self.y
 
-        self.color_palette = color_palette
+        # if tuple, typecasting color_palette to a list
+        self.color_palette = list(color_palette)
         self.aggregate_fn = aggregate_fn
         self.width = width
         self.height = height